feat(PF-18): TEST, MMMMMMMMMMMMM

+++ src/users/users.service.spec.ts
@@ -12,22 +12,23 @@ import { timeString } from '../utils/time';
 import { UpdateUserDto } from './dto/update-user.dto';
 
 describe('The UserService', () => {
-    // let userService = new UsersService(
-    //     new User() as any,
-    //     new RolesService(Role, UserRoles));
+
     let userService: UsersService
     let memDb: Sequelize;
     beforeAll(async () => {
         memDb = await createMemDB([User, Role, UserRoles, Post, Blog])
         userService = new UsersService(User, new RolesService(Role, UserRoles))
     });
-    afterAll(() => memDb.close());
+
+        afterAll(async () => {
+        const users = await User.findAll({where: { firstName: 'test'}})
+
+        for (const user of users) {
+            await user.destroy();
+        }
+        memDb.close()
+    });
     describe('Testing UsersService', () => {
-        it('should return all users', async () => {
-            const users = await userService.getAllUsers()
-            expect(users).toBeDefined();
-            expect(Array.isArray(users)).toBe(true)
-        })
 
         it('should create user', async () => {
             const mockEmail = `test${timeString}@example.com`
@@ -37,24 +38,35 @@ describe('The UserService', () => {
             expect(createdUser.email).toBe(mockEmail);
             expect(createdUser.firstName).toBe(dto.firstName);
             expect(createdUser.lastName).toBe(dto.lastName);
-        })
+        }, 10000)
+
+        it('should return all users', async () => {
+            const users = await userService.getAllUsers()
+            expect(users).toBeDefined();
+            expect(Array.isArray(users)).toBe(true)
+        }, 10000)
 
         it('should update user', async () => {
             const user = await User.findOne({where: {firstName: "test"}})
-            const mockEmail = `test${timeString}@example.com`
-            const dto: UpdateUserDto = { email: mockEmail, password: "qwerty123", firstName: "test", lastName: "updated"}
-            const updatedUser = await userService.update(dto, user.id, user.id, ["Moderator"])
-            expect(updatedUser).toBeDefined();
-            expect(updatedUser.email).toBe(mockEmail);
-            expect(updatedUser.firstName).toBe(dto.firstName);
-            expect(updatedUser.lastName).toBe(dto.lastName);
-        })
+            if(user){
+                const mockEmail = `test${timeString}@example.com`
+                const dto: UpdateUserDto = { email: mockEmail, password: "qwerty123", firstName: "test", lastName: "updated"}
+                const updatedUser = await userService.update(dto, user.id, user.id, ["Moderator"])
+                expect(updatedUser).toBeDefined();
+                expect(updatedUser.email).toBe(mockEmail);
+                expect(updatedUser.firstName).toBe(dto.firstName);
+                expect(updatedUser.lastName).toBe(dto.lastName);
+            }
+
+        }, 10000)
 
         it('should delete user', async () => {
             const user = await User.findOne({where: {firstName: "test"}})
-            await userService.delete(user.id, 1)
-            const deletedUser = await User.findOne({ where: { id: user.id } });
-            expect(deletedUser).toBeNull();
-        })
+            if(user){
+                await userService.delete(user.id, 1)
+                const deletedUser = await User.findOne({ where: { id: user.id } });
+                expect(deletedUser).toBeNull();
+             }
+        }, 10000)
     })
 });
